% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calandarHeat}
\alias{calandarHeat}
\alias{calendarHeat}
\title{Calandar Heatmap}
\usage{
calendarHeat(dates, values, minv, maxv, colors, ncolors = 99, title,
  date.form = "\%Y-\%m-\%d", l = T)
}
\arguments{
\item{dates}{a vector of dates}

\item{values}{a vector of values}

\item{minv}{min value}

\item{maxv}{max value}

\item{colors}{a vector of colors}

\item{ncolors}{number of colors}

\item{title}{title of the calandar heatmap}

\item{date.form}{date format}

\item{l}{TRUE: plot legend}
}
\description{
Plot a calandar heatmap
}
\examples{
days <- seq(as.Date("2016-01-1"), as.Date("2016-12-31"), "days")
values <- runif(length(days))
calendarHeat(days, values)
}
